*** empty log message ***

+++ StRoot/macros/embedding/daqtags/getruns.sh
@@ -1,5 +1,18 @@
 #!/bin/bash
 
+if [[ ! $HOST =~ "rcas" ]] ; then
+   echo this code can only be used for RCF!
+   exit
+fi
+
+if [[ ! $# -eq 2 ]] ; then
+   echo ""
+   echo " Usage : $0 [good_daqfilelist.daq.list] [#files] | sort | uniq | wc -l "
+   echo " output filelist : good_daqfilelist.[#files].daq.list. "
+   echo ""
+   exit
+fi
+
 list=`basename $1 .daq.list`.${2}.daq.list
 rm -f $list
